+++ subprojects/core/src/test/groovy/org/gradle/api/internal/DefaultDomainObjectCollectionTest.groovy
@@ -378,137 +378,6 @@ class DefaultDomainObjectCollectionTest extends AbstractDomainObjectCollectionSp
         container.add("a")
     }
 
-    def callsVetoActionBeforeObjectIsAdded() {
-        def action = Mock(Action)
-        container.beforeChange(action)
-
-        when:
-        container.add("a")
-
-        then:
-        1 * action.execute(null)
-        0 * _
-    }
-
-    def objectIsNotAddedWhenVetoActionThrowsAnException() {
-        def action = Mock(Action)
-        def failure = new RuntimeException()
-        container.beforeChange(action)
-
-        when:
-        container.add("a")
-
-        then:
-        def e = thrown(RuntimeException)
-        e == failure
-
-        and:
-        1 * action.execute(null) >> { throw failure }
-
-        and:
-        !toList(container).contains("a")
-    }
-
-    def callsVetoActionOnceBeforeCollectionIsAdded() {
-        def action = Mock(Action)
-        container.beforeChange(action)
-
-        when:
-        container.addAll(["a", "b"])
-
-        then:
-        1 * action.execute(null)
-        0 * _
-    }
-
-    def callsVetoActionBeforeObjectIsRemoved() {
-        def action = Mock(Action)
-        container.beforeChange(action)
-
-        when:
-        container.remove("a")
-
-        then:
-        1 * action.execute(null)
-        0 * _
-    }
-
-    def callsVetoActionBeforeObjectIsRemovedUsingIterator() {
-        def action = Mock(Action)
-
-        container.add("a")
-        container.beforeChange(action)
-
-        def iterator = container.iterator()
-        iterator.next()
-
-        when:
-        iterator.remove()
-
-        then:
-        1 * action.execute(null)
-        0 * _
-    }
-
-    def objectIsNotRemovedWhenVetoActionThrowsAnException() {
-        def action = Mock(Action)
-        def failure = new RuntimeException()
-
-        container.add("a")
-        container.beforeChange(action)
-
-        when:
-        container.remove("a")
-
-        then:
-        def e = thrown(RuntimeException)
-        e == failure
-
-        and:
-        1 * action.execute(null) >> { throw failure }
-
-        and:
-        toList(container).contains("a")
-    }
-
-    def callsVetoActionBeforeCollectionIsCleared() {
-        def action = Mock(Action)
-        container.beforeChange(action)
-
-        when:
-        container.clear()
-
-        then:
-        1 * action.execute(null)
-        0 * _
-    }
-
-    def callsVetoActionOnceBeforeCollectionIsRemoved() {
-        def action = Mock(Action)
-        container.beforeChange(action)
-
-        when:
-        container.removeAll(["a", "b"])
-
-        then:
-        1 * action.execute(null)
-        0 * _
-    }
-
-    def callsVetoActionOnceBeforeCollectionIsIntersected() {
-        def action = Mock(Action)
-        container.add("a")
-        container.add("b")
-        container.beforeChange(action)
-
-        when:
-        container.retainAll(toList())
-
-        then:
-        1 * action.execute(null)
-        0 * _
-    }
-
     def canRemoveAndMaintainOrder() {
         container.add("b")
         container.add("a")
